requirements.txt
fastapi
uvicorn[standard]
pillow
websocket-client
python-vlc
requests
